+++ packages/assets-controllers/src/TokenRatesController.test.ts
@@ -213,7 +213,7 @@ describe('TokenRatesController', () => {
   });
 
   it('should update all rates', async () => {
-    new NetworkController({ messenger });
+    new NetworkController({ messenger, trackMetaMetricsEvent: jest.fn() });
     const preferences = new PreferencesController();
     const tokensController = new TokensController({
       onPreferencesStateChange: (listener) => preferences.subscribe(listener),