+++ packages/transaction-controller/src/TransactionController.test.ts
@@ -135,9 +135,10 @@ const MOCK_NETWORK = {
     isCustomNetwork: false,
     networkDetails: { isEIP1559Compatible: false },
     providerConfig: {
-      type: 'goerli' as NetworkType,
+      type: NetworkType.goerli,
       chainId: NetworksChainId.goerli,
     },
+    networkConfigurations: {},
   },
   subscribe: () => undefined,
 };
@@ -148,16 +149,21 @@ const MOCK_NETWORK_CUSTOM = {
     isCustomNetwork: true,
     networkDetails: { isEIP1559Compatible: false },
     providerConfig: {
-      type: 'rpc' as NetworkType,
+      type: NetworkType.rpc,
       chainId: '10',
     },
+    networkConfigurations: {},
   },
   subscribe: () => undefined,
 };
 const MOCK_NETWORK_WITHOUT_CHAIN_ID = {
   getProvider: () => PROVIDER,
   isCustomNetwork: false,
-  state: { network: '5', providerConfig: { type: 'goerli' as NetworkType } },
+  state: {
+    network: '5',
+    providerConfig: { type: NetworkType.goerli },
+    networkConfigurations: {},
+  },
   subscribe: () => undefined,
 };
 const MOCK_MAINNET_NETWORK = {
@@ -167,9 +173,10 @@ const MOCK_MAINNET_NETWORK = {
     isCustomNetwork: false,
     networkDetails: { isEIP1559Compatible: false },
     providerConfig: {
-      type: 'mainnet' as NetworkType,
+      type: NetworkType.mainnet,
       chainId: NetworksChainId.mainnet,
     },
+    networkConfigurations: {},
   },
   subscribe: () => undefined,
 };
@@ -180,9 +187,10 @@ const MOCK_CUSTOM_NETWORK = {
     isCustomNetwork: true,
     networkDetails: { isEIP1559Compatible: false },
     providerConfig: {
-      type: 'rpc' as NetworkType,
+      type: NetworkType.rpc,
       chainId: '80001',
     },
+    networkConfigurations: {},
   },
   subscribe: () => undefined,
 };